+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
-
+### Navigator
+* Add a command to revert selected items to the previous version,
+so you can fix you mistake if you accidentally miss clicked 
 
 # 4.22.0